♻️ refactor: middleware chain
- 재귀적으로 미들웨어를 실행하는 chain 함수 설정
- 기존 middleware matcher 에 '/' 경로 추가
